# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Corp chat API","version":"1.0.0"},"paths":{"/user/register":{"post":{"tags":["User methods"],"summary":"Register User","operationId":"register_user_user_register_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInfo"}}}}}}},"/user/{user_id}":{"get":{"tags":["User methods"],"summary":"Get User Info","operationId":"get_user_info_user__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/UserInfo"},{"type":"null"}],"title":"Response Get User Info User  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/updateSettings/{user_id}":{"put":{"tags":["User methods"],"summary":"Get User Info","operationId":"get_user_info_user_updateSettings__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserSettings"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/config/getByUserId/{user_id}":{"get":{"tags":["User methods"],"summary":"Get User Info","operationId":"get_user_info_config_getByUserId__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Config"},"title":"Response Get User Info Config Getbyuserid  User Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/config/getTypes":{"get":{"tags":["User methods"],"summary":"Get Config Types","operationId":"get_config_types_config_getTypes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ConfigType"},"type":"array","title":"Response Get Config Types Config Gettypes Get"}}}}}}},"/config/create":{"post":{"tags":["User methods"],"summary":"Create Config","operationId":"create_config_config_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/config/update/{config_id}":{"put":{"tags":["User methods"],"summary":"Update Config","operationId":"update_config_config_update__config_id__put","parameters":[{"name":"config_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Config Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfigUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Config"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/create/{user_id}":{"post":{"tags":["Chat methods"],"summary":"Create Chat","operationId":"create_chat_chat_create__user_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Chat"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/{chat_id}/generate":{"post":{"tags":["Chat methods"],"summary":"Generate Message","operationId":"generate_message_chat__chat_id__generate_post","parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Chat Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMessage"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/{chat_id}/messages":{"get":{"tags":["Chat methods"],"summary":"Get Messages","operationId":"get_messages_chat__chat_id__messages_get","parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Chat Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"},"title":"Response Get Messages Chat  Chat Id  Messages Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/create/{user_id}":{"post":{"tags":["Rag methods"],"summary":"Create Rag","operationId":"create_rag_rag_create__user_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_rag_rag_create__user_id__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rag"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/delete/{rag_id}":{"delete":{"tags":["Rag methods"],"summary":"Delete Rag","operationId":"delete_rag_rag_delete__rag_id__delete","parameters":[{"name":"rag_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Rag Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/rag/getByUser_id/{user_id}":{"get":{"tags":["Rag methods"],"summary":"Get Rag By User Id","operationId":"get_rag_by_user_id_rag_getByUser_id__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_create_rag_rag_create__user_id__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_create_rag_rag_create__user_id__post"},"Chat":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","created_at"],"title":"Chat"},"Config":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"type_id":{"type":"string","format":"uuid","title":"Type Id"},"url":{"type":"string","title":"Url"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"llm_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Llm Name"}},"type":"object","required":["id","user_id","type_id","url"],"title":"Config"},"ConfigCreate":{"properties":{"user_id":{"type":"string","format":"uuid","title":"User Id"},"type_id":{"type":"string","format":"uuid","title":"Type Id"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"llm_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Llm Name"}},"type":"object","required":["user_id","type_id"],"title":"ConfigCreate"},"ConfigType":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"ConfigType"},"ConfigUpdate":{"properties":{"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"llm_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Llm Name"}},"type":"object","title":"ConfigUpdate"},"GenerateMessage":{"properties":{"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"GenerateMessage"},"GenerateResponse":{"properties":{"userMessage":{"$ref":"#/components/schemas/MessageResponse"},"assistantMessage":{"$ref":"#/components/schemas/MessageResponse"}},"type":"object","required":["userMessage","assistantMessage"],"title":"GenerateResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","role","content","created_at"],"title":"MessageResponse"},"Rag":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"file_name":{"type":"string","title":"File Name"}},"type":"object","required":["id","file_name"],"title":"Rag"},"UpdateUserSettings":{"properties":{"system_msg":{"type":"string","title":"System Msg"},"msg_count":{"type":"integer","title":"Msg Count"}},"type":"object","required":["system_msg","msg_count"],"title":"UpdateUserSettings"},"User":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"config_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Config Id"},"system_msg":{"type":"string","title":"System Msg"},"msg_count":{"type":"integer","title":"Msg Count"}},"type":"object","required":["id","config_id","system_msg","msg_count"],"title":"User"},"UserInfo":{"properties":{"chats":{"items":{"$ref":"#/components/schemas/Chat"},"type":"array","title":"Chats"},"config":{"anyOf":[{"$ref":"#/components/schemas/Config"},{"type":"null"}]},"user":{"$ref":"#/components/schemas/User"}},"type":"object","required":["chats","config","user"],"title":"UserInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}